# backend/go/src/app/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Game Master Notes Backend API",
    "version": "0.1.0",
    "description": "HTTP API for Game Master Notes domain services."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Default local server (runtime metadata is dynamically injected)."
    }
  ],
...
  "paths": {
...
    "/worlds": {
      "get": {
        "tags": ["Worlds"],
        "summary": "List worlds",
        "parameters": [
          { "$ref": "#/components/parameters/Offset" },
          { "$ref": "#/components/parameters/Limit" },
          { "$ref": "#/components/parameters/IncludeDeleted" }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ListWorldsResponse" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Worlds"],
        "summary": "Create world",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateWorldRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/WorldResponse" }
              }
            }
          }
        }
      }
    },
    "/worlds/{id}": {
      "get": {
        "tags": ["Worlds"],
        "summary": "Get world by id",
        "parameters": [
          { "$ref": "#/components/parameters/ID" },
          { "$ref": "#/components/parameters/IncludeDeleted" }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/WorldResponse" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Worlds"],
        "summary": "Update world",
        "parameters": [{ "$ref": "#/components/parameters/ID" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateWorldRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/WorldResponse" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Worlds"],
        "summary": "Delete world",
        "parameters": [{ "$ref": "#/components/parameters/ID" }],
        "responses": { "204": { "description": "No Content" } }
      }
    },
...
    "/tags": {
      "get": {
        "tags": ["Tags"],
        "summary": "List tags",
        "parameters": [
          { "$ref": "#/components/parameters/Offset" },
          { "$ref": "#/components/parameters/Limit" },
          { "$ref": "#/components/parameters/IncludeDeleted" }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ListTagsResponse" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Tags"],
        "summary": "Create tag",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateTagRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TagResponse" }
              }
            }
          }
        }
      }
    },
    "/tags/{id}": {
      "get": {
        "tags": ["Tags"],
        "summary": "Get tag by id",
        "parameters": [
          { "$ref": "#/components/parameters/ID" },
          { "$ref": "#/components/parameters/IncludeDeleted" }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TagResponse" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Tags"],
        "summary": "Update tag",
        "parameters": [{ "$ref": "#/components/parameters/ID" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateTagRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TagResponse" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Tags"],
        "summary": "Delete tag",
        "parameters": [{ "$ref": "#/components/parameters/ID" }],
        "responses": { "204": { "description": "No Content" } }
      }
    },
...
  },
  "components": {
    "parameters": {
      "ID": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "Resource identifier (ULID string)."
      },
...
      "Offset": {
        "name": "offset",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "format": "int32",
          "minimum": 0,
          "default": 0
        }
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "format": "int32",
          "minimum": 1,
          "default": 20
        }
      },
      "IncludeDeleted": {
        "name": "include_deleted",
        "in": "query",
        "required": false,
        "schema": {
          "type": "boolean",
          "default": false
        }
      },
...
    },
    "schemas": {
...
      "CreateWorldRequest": {
        "type": "object",
        "properties": {
          "plane_id": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "status": { "type": "string", "enum": ["draft", "active", "archived"] }
        }
      },
      "UpdateWorldRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" },
          "status": { "type": "string", "enum": ["draft", "active", "archived"] }
        }
      },
      "WorldResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "plane_id": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "status": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" },
          "deleted_at": { "type": "string", "format": "date-time", "nullable": true },
          "version": { "type": "integer", "format": "int32" }
        }
      },
      "ListWorldsResponse": {
        "type": "object",
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/WorldResponse" } },
          "offset": { "type": "integer", "format": "int32" },
          "limit": { "type": "integer", "format": "int32" },
          "include_deleted": { "type": "boolean" }
        }
      },
...
      "CreateTagRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "campaign_id": { "type": "string", "nullable": true }
        }
      },
      "UpdateTagRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "campaign_id": { "type": "string", "nullable": true }
        }
      },
      "TagResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "campaign_id": { "type": "string", "nullable": true },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" },
          "deleted_at": { "type": "string", "format": "date-time", "nullable": true },
          "version": { "type": "integer", "format": "int32" }
        }
      },
      "ListTagsResponse": {
        "type": "object",
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/TagResponse" } },
          "offset": { "type": "integer", "format": "int32" },
          "limit": { "type": "integer", "format": "int32" },
          "include_deleted": { "type": "boolean" }
        }
      },
...
    }
  }
}